done tidying dist tests
