+++ starsim/distributions.py
@@ -37,10 +37,9 @@ def find_dists(obj, verbose=False):
 class Dists:
     """ Class for managing a collection of Dist objects """
 
-    def __init__(self, *args, base_seed=None, sim=None):
-        args = list(args)
-        if len(args) == 1: args = args[0]
-        self.obj = args
+    def __init__(self, obj=None, *args, base_seed=None, sim=None):
+        if len(args): obj = [obj] + list(args)
+        self.obj = obj
         self.dists = None
         self.base_seed = base_seed
         self.sim = sim
